tests/wpt/tests/navigation-api/scroll-behavior/scroll-without-intercept.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<div style="height: 1000px; width: 1000px;"></div>
<div id="frag"></div>
<script>
promise_test(async t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  await new Promise(resolve => window.onload = () => t.step_timeout(resolve, 0));

  navigation.addEventListener("navigate", t.step_func(e => {
    assert_throws_dom("InvalidStateError", () => e.scroll());
  }), { once : true });
  await navigation.navigate("#frag").finished;
}, "scroll: scroll() should throw for non-intercept");
</script>
</body>
